Update README.txt
Insignificant change, to allow me to contact you...

This is a teeny tiny library, but it is used in downstream projects. Generally speaking, now that the maintainer has abandoned it,  I don't expect it to be too disruptive for all the downstream projects either internalize this code or replace it with another library. 

However, I do like how small and simple this version is. So basically out of that sentiment, I'd like to see it updated on pypi with the fixes you've patched in. 

Since your patch is better one, I thought it best to allow you to do that if you'd like.

https://www.python.org/dev/peps/pep-0541/#how-to-request-a-name-transfer

I'd be happy to do this if you don't have the time.